Removed unnecessarily pesky uint64 type from unixtimes

+++ drivers/reddit/types/types.go
@@ -10,8 +10,8 @@ type RedditPost struct {
 	Id            string
 	Name          string // Note that this more of an ID, used in the "after" parameter of the scraper
 	IsPublished   bool
-	TimeCreated   uint64 `mapstructure:"time_created"`
-	TimeStored    uint64 `mapstructure:"time_stored"`
+	TimeCreated   int64 `mapstructure:"time_created"`
+	TimeStored    int64 `mapstructure:"time_stored"`
 	Permalink     string
 	IsActive      bool
 	IsSticky      bool
@@ -29,8 +29,8 @@ func NewRedditPostFromBotPost(bp *reddit.Post) (p RedditPost) {
 	p.Name = bp.Name
 	p.Score = int64(bp.Score)
 	p.IsPublished = false
-	p.TimeCreated = bp.CreatedUTC
-	p.TimeStored = bp.CreatedUTC
+	p.TimeCreated = int64(bp.CreatedUTC)
+	p.TimeStored = int64(bp.CreatedUTC)
 	p.Permalink = bp.Permalink
 	p.IsActive = !bp.Deleted
 	p.IsSticky = bp.Stickied
